+++ tests/Microsoft.Bot.Builder.Tests/StateActivity_InterationTests.cs
@@ -11,13 +11,13 @@ using System.Linq;
 namespace Microsoft.Bot.Builder.Tests
 {
     [TestClass]
-    class StateActivity_InterationTests
+    [TestCategory("State Management")]
+    public class StateActivity_InterationTests
     {
-
         [TestMethod]
-        public void State_Conversation_DoNOTRememberConversationAfterEndOfConversation()
+        public async Task State_Conversation_DoNOTRememberConversationAfterEndOfConversation()
         {
-            IList<IActivity> testActivities = InteractionTestsActivityHelper.StartNewConversation();
+            IList<IActivity> testActivities = StateActivity_Helper.StartNewConversation();
 
             TestAdapter adapter = new TestAdapter()
                 .Use(new UserStateManagerMiddleware(new MemoryStorage()))
@@ -26,13 +26,19 @@ namespace Microsoft.Bot.Builder.Tests
             var testFlow = new TestFlow(adapter, async (context) =>
             {
                 Assert.IsNotNull(context.State, "context.state should exist");
-                string text = context.Request.AsMessageActivity()?.Text.Trim().ToLower();
-                if (text != null)
+                switch (context.Request.Type)
                 {
-                    var parts = text.Split(' ', StringSplitOptions.RemoveEmptyEntries);
-                    var name = parts?[1] ?? "value";
-                    var input = parts?[2] ?? "test";
-                    switch (parts[0])
+                    case ActivityTypes.ConversationUpdate:
+                    case ActivityTypes.EndOfConversation:
+                    case ActivityTypes.DeleteUserData:
+                    case ActivityTypes.Message:
+                    default:
+                        break;
+                }
+                string text = context.Request.AsMessageActivity()?.Text?.Trim().ToLower();
+                    var name = "value";
+                    var input = "test";
+                    switch (text)
                     {
                         case "set":
                             context.State.ConversationProperties[name] = input;
@@ -42,22 +48,23 @@ namespace Microsoft.Bot.Builder.Tests
                             string state = context.State.ConversationProperties[name];
                             context.Reply(state);
                             break;
-                    }
+                    default:
+                        break;
                 }
             });
-            foreach (var activity in InteractionTestsActivityHelper.StartNewConversation())
+            foreach (var activity in StateActivity_Helper.StartNewConversation())
             {
                 testFlow.Send(activity);
             }
-            //await testFlow
-            //    .Test("get", string.Empty, "Conversation state should NOT be remembered after EndOfConversation.")
-            //    .StartTest();
+            await testFlow
+                .Test("get", string.Empty, "Conversation state should NOT be remembered after EndOfConversation.")
+                .StartTest();
         }
 
         [TestMethod]
-        public async void State_User_RememberUserAfterEndOfConversation()
+        public async Task State_User_RememberUserAfterEndOfConversation()
         {
-            IList<IActivity> testActivities = InteractionTestsActivityHelper.StartNewConversation();
+            IList<IActivity> testActivities = StateActivity_Helper.StartNewConversation();
 
             TestAdapter adapter = new TestAdapter()
                 .Use(new UserStateManagerMiddleware(new MemoryStorage()))
@@ -69,10 +76,9 @@ namespace Microsoft.Bot.Builder.Tests
                 string text = context.Request.AsMessageActivity()?.Text.Trim().ToLower();
                 if (text != null)
                 {
-                    var parts = text.Split(' ', StringSplitOptions.RemoveEmptyEntries);
-                    var name = parts?[1] ?? "value";
-                    var input = parts?[2] ?? "test";
-                    switch (parts[0])
+                    var name = "value";
+                    var input = "test";
+                    switch (text)
                     {
                         case "set":
                             context.State.UserProperties[name] = input;
@@ -85,7 +91,7 @@ namespace Microsoft.Bot.Builder.Tests
                     }
                 }
             });
-            foreach (var activity in InteractionTestsActivityHelper.StartNewConversation())
+            foreach (var activity in StateActivity_Helper.StartNewConversation())
             {
                 testFlow.Send(activity);
             }
@@ -95,9 +101,9 @@ namespace Microsoft.Bot.Builder.Tests
         }
 
         [TestMethod]
-        public async void State_User_DoNOTRememberUserAfterDeleteUserData()
+        public async Task State_User_DoNOTRememberUserAfterDeleteUserData()
         {
-            IList<IActivity> testActivities = InteractionTestsActivityHelper.StartNewConversation();
+            IList<IActivity> testActivities = StateActivity_Helper.StartNewConversation();
 
             TestAdapter adapter = new TestAdapter()
                 .Use(new UserStateManagerMiddleware(new MemoryStorage()))
@@ -109,10 +115,9 @@ namespace Microsoft.Bot.Builder.Tests
                 string text = context.Request.AsMessageActivity()?.Text.Trim().ToLower();
                 if (text != null)
                 {
-                    var parts = text.Split(' ', StringSplitOptions.RemoveEmptyEntries);
-                    var name = parts?[1] ?? "value";
-                    var input = parts?[2] ?? "test";
-                    switch (parts[0])
+                    var name = "value";
+                    var input = "test";
+                    switch (text)
                     {
                         case "set":
                             context.State.UserProperties[name] = input;
@@ -125,13 +130,124 @@ namespace Microsoft.Bot.Builder.Tests
                     }
                 }
             });
-            foreach (var activity in InteractionTestsActivityHelper.StartNewConversation())
+            foreach (var activity in StateActivity_Helper.StartNewConversation())
             {
                 testFlow.Send(activity);
             }
             await testFlow
                 .Test("get", string.Empty, "User state should NOT be remembered after DeleteUserData.")
                 .StartTest();
+        }
+    }
+
+    public static class StateActivity_Helper
+    {
+        public static ConversationReference CreateConversationReference(
+            string userName = "User1", string conversationName = "Convo1")
+        {
+            var conversationReference = new ConversationReference
+            {
+                ChannelId = "test",
+                ServiceUrl = "https://test.com"
+            };
+            conversationReference.User = new ChannelAccount(userName.ToLower(), userName);
+            conversationReference.Bot = new ChannelAccount("bot", "Bot");
+            conversationReference.Conversation = new ConversationAccount(
+                false, conversationName.ToLower(), conversationName);
+            return conversationReference;
+        }
+
+        public static IEnumerable<IActivity> StartConversation(ConversationReference reference)
+        {
+            yield return new Activity()
+            {
+                Type = ActivityTypes.ConversationUpdate,
+                Conversation = reference.Conversation,
+                From = reference.User,
+                Recipient = reference.Bot,
+                ServiceUrl = reference.ServiceUrl,
+                MembersAdded = new List<ChannelAccount> { reference.User },
+                MembersRemoved = new List<ChannelAccount>()
+            };
+            yield return new Activity()
+            {
+                Type = ActivityTypes.ConversationUpdate,
+                Conversation = reference.Conversation,
+                From = reference.User,
+                Recipient = reference.Bot,
+                ServiceUrl = reference.ServiceUrl,
+                MembersAdded = new List<ChannelAccount> { reference.Bot },
+                MembersRemoved = new List<ChannelAccount>()
+            };
+        }
+
+        public static IActivity CreateMessageActivity(ConversationReference reference, string message)
+        {
+            return new Activity()
+            {
+                Type = ActivityTypes.Message,
+                Conversation = reference.Conversation,
+                From = reference.User,
+                Recipient = reference.Bot,
+                ServiceUrl = reference.ServiceUrl,
+                Text = message
+            };
+        }
+
+        public static IActivity CreateEndOfConversation(ConversationReference reference, string endOfConversationCode = EndOfConversationCodes.Unknown)
+        {
+            return new Activity()
+            {
+                Type = ActivityTypes.EndOfConversation,
+                Conversation = reference.Conversation,
+                From = reference.User,
+                Recipient = reference.Bot,
+                ServiceUrl = reference.ServiceUrl,
+                Code = endOfConversationCode
+            };
+        }
+
+        public static IActivity CreateDeleteUserData(ConversationReference reference)
+        {
+            return new Activity()
+            {
+                Type = ActivityTypes.DeleteUserData,
+                Conversation = reference.Conversation,
+                From = reference.User,
+                Recipient = reference.Bot,
+                ServiceUrl = reference.ServiceUrl
+            };
+        }
+
+        public static IList<IActivity> StartNewConversation()
+        {
+            var convo1 = CreateConversationReference(conversationName: "Convo1");
+            var convo2 = CreateConversationReference(conversationName: "Convo2");
+            var userAccount = convo1.User;
+            var botAccount = convo1.Bot;
+
+            var activities = new List<IActivity>();
+            activities.AddRange(StartConversation(convo1));
+            activities.Add(CreateMessageActivity(convo1, "set"));
+            activities.Add(CreateEndOfConversation(convo1));
+            activities.AddRange(StartConversation(convo2));
+
+            return activities;
+        }
+
+        public static IList<IActivity> SetUpThenSendDeleteUserData()
+        {
+            var convo1 = CreateConversationReference(conversationName: "Convo1");
+            var convo2 = CreateConversationReference(conversationName: "Convo2");
+            var userAccount = convo1.User;
+            var botAccount = convo1.Bot;
+
+            var activities = new List<IActivity>();
+            activities.AddRange(StartConversation(convo1));
+            activities.Add(CreateMessageActivity(convo1, "set"));
+            activities.Add(CreateDeleteUserData(convo1));
+
+            return activities;
         }
     }
 }
